Separated the game logic to another file.
